gms-common/python/master-coi-data-client/requirements.txt
pandas
requests
termcolor
matplotlib
numpy
python-dateutil
pyyaml
